+++ lib/private.example.dart
@@ -5,4 +5,6 @@ class Privates {
 
   static final String mongoDbUri =
       "mongodb://<username>:<password>@<ip>:<port>/<database-name>?authSource=admin&directConnection=true";
+
+  static final String encryptionKey = "add 32 digit encryption key here";
 }